+++ CHANGELOG.rst
@@ -92,6 +92,19 @@ of the list).
 
 - Modified the check on the tweakwcs_output to be more robust[#260]
 
+- Ensure singletons do not use the match_relative_fit algorithm [#259]
+
+- Entirely re-wrote check_and_get_data function in alignimages.py.
+  Updates to the overall hlapipeline alignment code and updates/clean-up 
+  to the associated regression tests migrated from the original hlapipeline
+  package [#283].
+
+- Fixed improper logic caused by nested if statements in analyze.py found
+  when testing [#283].
+
+- Fixed a logic bug in alignimages.py to force the run_align() function
+  to always return the filtered table contents. [#302]
+
 
 2.2.6 (02-Nov-2018)
 ===================